+++ app/src/main/java/womenproject/com/mybury/presentation/dialog/LoadFailDialog.kt
@@ -3,7 +3,7 @@ package womenproject.com.mybury.presentation.dialog
 import android.view.View
 import womenproject.com.mybury.presentation.base.BaseNormalDialogFragment
 
-class LoadFailDialog(val onConfirmClicked: () -> Unit) : BaseNormalDialogFragment() {
+class LoadFailDialog(val onConfirmClicked: (() -> Unit)? = null) : BaseNormalDialogFragment() {
 
     init {
         TITLE_MSG = "데이터 로드 실패"
@@ -16,7 +16,9 @@ class LoadFailDialog(val onConfirmClicked: () -> Unit) : BaseNormalDialogFragmen
 
     override fun createOnClickConfirmListener(): View.OnClickListener {
         return View.OnClickListener {
-            onConfirmClicked()
+            onConfirmClicked?.let {
+                it.invoke()
+            }
             dismiss()
         }
     }
