+++ app/src/main/java/womenproject/com/mybury/presentation/viewmodels/CategoryViewModel.kt
@@ -1,21 +1,26 @@
 package womenproject.com.mybury.presentation.viewmodels
 
+import android.annotation.SuppressLint
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.viewModelScope
-import com.zinc.domain.usecase.category.LoadCategoryListUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
+import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.schedulers.Schedulers
 import kotlinx.coroutines.launch
-import womenproject.com.mybury.data.Category
-import womenproject.com.mybury.data.toCategoryData
+import womanproject.com.mybury.domain.usecase.category.LoadCategoryListUseCase
+import womenproject.com.mybury.data.*
+import womenproject.com.mybury.data.model.LoadState
+import womenproject.com.mybury.data.network.apiInterface
 import womenproject.com.mybury.presentation.base.BaseViewModel
 import javax.inject.Inject
+
 
 @HiltViewModel
 class CategoryViewModel @Inject constructor(
     private val loadCategoryListUseCase: LoadCategoryListUseCase
-) : BaseViewModel() {
+): BaseViewModel() {
 
     private val _categoryLoadState = MutableLiveData<LoadState>()
     val categoryLoadState: LiveData<LoadState> = _categoryLoadState
@@ -23,7 +28,7 @@ class CategoryViewModel @Inject constructor(
     private val _categoryList = MutableLiveData<List<Category>>()
     val categoryList: LiveData<List<Category>> = _categoryList
 
-    fun getCategoryList() {
+    fun loadCategoryList() {
         if (accessToken == null || userId == null) {
             _categoryLoadState.value = LoadState.START
             return
@@ -50,4 +55,128 @@ class CategoryViewModel @Inject constructor(
             }
         }
     }
+
+    @SuppressLint("CheckResult")
+    fun removeCategoryItem(categoryId: HashSet<String>, callBack: Simple3CallBack) {
+        if(accessToken==null || userId == null) {
+            callBack.fail()
+            return
+        }
+
+        // 카테고리 제거
+        val list = ArrayList<String>()
+        for(i in categoryId) {
+            list.add(i)
+        }
+        val request = RemoveCategoryRequest(userId, list)
+        apiInterface.removeCategoryItem(accessToken, request)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    when {
+                        it.retcode == "200" -> callBack.success()
+                        it.retcode == "301" -> getRefreshToken(object : SimpleCallBack {
+                            override fun success() {
+                                callBack.restart()
+                            }
+
+                            override fun fail() {
+                                callBack.fail()
+                            }
+                        })
+                        else -> callBack.fail()
+                    }
+                }) {
+                    callBack.fail()
+                }
+
+    }
+
+    @SuppressLint("CheckResult")
+    fun addCategoryItem(categoryName: String, callBack: Simple3CallBack) {
+        if(accessToken==null || userId == null) {
+            callBack.fail()
+            return
+        }
+
+        callBack.start()
+        val request = AddCategoryRequest(userId, categoryName)
+        apiInterface.addNewCategoryItem(accessToken, request)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    when {
+                        it.retcode == "200" -> callBack.success()
+                        it.retcode == "301" -> getRefreshToken(object : SimpleCallBack {
+                            override fun success() {
+                                callBack.restart()
+                            }
+
+                            override fun fail() {
+                                callBack.fail()
+                            }
+                        })
+                        else -> callBack.fail()
+                    }
+                }) {
+                    callBack.fail()
+                }
+
+    }
+
+    @SuppressLint("CheckResult")
+    fun editCategoryItem(category : Category, categoryName: String, callBack: Simple3CallBack) {
+        if(accessToken==null || userId == null) {
+            callBack.fail()
+            return
+        }
+
+        callBack.start()
+
+        val request = EditCategoryNameRequest(userId, category.id, categoryName)
+        apiInterface.editCategoryItemName(accessToken, request)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    callBack.success()
+                }) {
+                   callBack.fail()
+                }
+    }
+
+    @SuppressLint("CheckResult")
+    fun changeCategoryStatus(list: List<Category>, callBack: Simple3CallBack) {
+        if(accessToken==null || userId == null) {
+            callBack.fail()
+            return
+        }
+
+        var idList = arrayListOf<String>()
+        for(i in list) {
+            idList.add(i.id)
+        }
+        val request = ChangeCategoryStatusRequest(userId, idList)
+        callBack.start()
+        apiInterface.changeCategoryList(accessToken, request)
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread())
+                .subscribe({
+                    when (it.retcode) {
+                        "200" -> callBack.success()
+                        "301" -> getRefreshToken(object : SimpleCallBack {
+                            override fun success() {
+                                callBack.restart()
+                            }
+
+                            override fun fail() {
+                                callBack.fail()
+                            }
+                        })
+                        else -> callBack.fail()
+                    }
+                }) {
+                    callBack.fail()
+                }
+    }
+
 }
