--- data/src/main/java/com/zinc/data/model/CategoryData.kt
@@ -1,17 +1,0 @@
-package com.zinc.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class BucketCategory(
-    val categoryList: List<DomainCategory>,
-    val retcode: String
-)
-
-@Serializable
-data class DomainCategory(
-    val name: String,
-    val id: String,
-    val priority: Int = 0,
-    val isDefault: String? = "N"
-)
