--- data/src/main/java/com/zinc/data/model/CommonData.kt
@@ -1,8 +1,0 @@
-package com.zinc.data.model
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SimpleResponse(
-    val retcode: String
-)
