+++ data/src/main/java/womenproject/com/mybury/data/model/CommonData.kt
@@ -1,0 +1,8 @@
+package womenproject.com.mybury.data.model
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class SimpleResponse(
+    val retcode: String
+)
