--- domain/src/main/java/com/zinc/domain/usecase/category/LoadCategoryListUseCase.kt
@@ -1,13 +1,0 @@
-package com.zinc.domain.usecase.category
-
-import com.zinc.data.model.BucketCategory
-import com.zinc.data.repository.MyBuryRepository
-import javax.inject.Inject
-
-class LoadCategoryListUseCase @Inject constructor(
-    private val myBuryRepository: MyBuryRepository
-) {
-    suspend fun invoke(token: String, userId: String): BucketCategory {
-        return myBuryRepository.loadCategoryList(token, userId)
-    }
-}
